requirements.txt
numpy==1.25.2
opencv-python==4.8.0.74
open3d==0.17.0
pillow==9.2.0
scipy==1.10.1
matplotlib==3.7.3
